Add data, initialise dvc
