+++ requirements.txt
@@ -8,3 +8,5 @@ google-cloud-storage
 prometheus-fastapi-instrumentator==5.7.1
 pandas>=1.1.5
 requests>=2.27.1
+dvc
+dvc-gs
